requirements.txt
absl-py==1.0.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
docker-pycreds==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.0
google-auth-oauthlib==0.4.6
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.2
Markdown==3.3.6
numpy==1.22.3
oauthlib==3.2.0
pathtools==0.1.2
Pillow==9.0.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
sentry-sdk==1.5.7
setproctitle==1.2.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
torch==1.10.2
torchvision==0.11.3
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.8
wandb==0.12.15
Werkzeug==2.0.3
yaspin==2.1.0
zipp==3.7.0
